feat: 添加部门接口功能

+++ backend/src/main/java/net/geekhour/loki/controller/AssetController.java
@@ -73,7 +73,7 @@ public class AssetController {
         }
         Integer offset = (pageIndex - 1) * pageSize;
         List<AssetDTO> assetList = assetService.getAssetList(offset, pageSize, name);
-        int total = assetService.countAssets(name);
+        Long total = assetService.countAssets(name);
         return ResponseEntity.ok(Map.of(
                 "code", 200,
                 "message", "success!",
